requirements.txt
agno
python-dotenv
openai



